# Optimization and fixes for conversion from `f64`