--- a/.github/workflows/ci.yml
+++ b/.github/workflows/ci.yml
@@ -54,5 +54,19 @@ jobs:
   run-example:
     runs-on: ubuntu-latest
     steps:
+      - uses: actions/checkout@v2
+      - run: cargo run --example usage --features i64
+
+  fuzz:
+    runs-on: ubuntu-latest
+    steps:
     - uses: actions/checkout@v2
-    - run: cargo run --example usage --features i64
+    - uses: actions-rs/toolchain@v1
+      with:
+        toolchain: nightly
+        profile: minimal
+        default: true
+        components: llvm-tools-preview
+    - run: cargo install cargo-fuzz
+    - run: cargo fuzz --version
+    - run: cargo fuzz run f64 -- -max_total_time=10s