--- a/CHANGELOG.md
+++ b/CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Fixed
+- Rare "from `f64`" bugs ([@quasiyoke](https://github.com/quasiyoke)).
 
 ## [0.6.1] - 2021-12-30
 ### Changed