--- a/Cargo.toml
+++ b/Cargo.toml
@@ -19,10 +19,10 @@ name = "fixed_point"
 path = "benches/fixed_point.rs"
 harness = false
 
-[[bench]]
-name = "serde"
-path = "benches/serde.rs"
-harness = false
+#[[bench]]
+#name = "serde"
+#path = "benches/serde.rs"
+#harness = false
 
 [[example]]
 name = "usage"
@@ -35,6 +35,7 @@ typenum = "1.12.0"
 derive_more = { version = "0.99.9", default-features = false }
 parity-scale-codec = { version = "2", default-features = false, features = ["derive"], optional = true }
 static_assertions = "1.1.0"
+rust_decimal = "1.22.0"
 
 [dev-dependencies]
 anyhow = { version = "1.0.38", default-features = false }